Refactored taskRouter.js and taskModel.js to work with auth layer so that user can only interact with their own tasks.

+++ models/taskModel.js
@@ -3,57 +3,61 @@ const { createClient } = require('@supabase/supabase-js');
 const supabase = createClient(process.env.SUPABASE_URL, process.env.SUPABASE_SERVICE_KEY)
 
 
-const getAllTasks = async (filterKey, filterValue) => {
+const getAllTasks = async (filterKey, filterValue, userId) => {
     if (filterKey && filterValue) {
         const { data, error } = await supabase
             .from('tasks')
             .select('*')
             .eq(filterKey, filterValue)
+            .eq('userId', userId)
 
         return { data, error }
     }
-    
-    const { data, error } = await supabase.from('tasks').select('*')
+
+    const { data, error } = await supabase.from('tasks').select('*').eq('userId', userId)
     return { data, error }
 };
 
 
-const getTaskByID = async (id) => {
+const getTaskByID = async (id, userId) => {
     const { data, error } = await supabase
         .from('tasks')
         .select('*')
         .eq('id', id)
+        .eq('userId', userId)
 
     return { data, error }
 }
 
 
-const insertTask = async (task) => {
+const insertTask = async (task, userId) => {
     const { data, error } = await supabase
         .from('tasks')
-        .insert(task)
+        .insert({ title: task.title, completed: task.completed, dueDate: task.dueDate, userId: userId })
         .select()
 
     return { data, error }
 }
 
 
-const updateTask = async (id, task) => {
+const updateTask = async (id, task, userId) => {
     const { data, error } = await supabase
         .from('tasks')
         .update(task)
         .eq('id', id)
+        .eq('userId', userId)
         .select()
 
     return { data, error }
 }
 
 
-const deleteTask = async (id) => {
+const deleteTask = async (id, userId) => {
     const { data, error } = await supabase
         .from('tasks')
         .delete()
         .eq('id', id)
+        .eq('userId', userId)
         .select()
 
     return { data, error }
